pyproject.toml
[project]
name = "squad44-status"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "MerreM", email = "merrem@0xdb.co.uk" }]
dependencies = ["requests>=2.32.3", "discord-py>=2.4.0", "python-dotenv>=1.0.1"]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = ["ruff>=0.6.5"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
main = "squad44_status.main"

[project.scripts]
squad44_status_bot = "squad44_status.__main__:entrypoint"

[tool.pdm]
distribution = true
